refactor(theme): split theme specs
