+++ src/framework/theme/component/theme/theme.spec.config.ts
@@ -1,0 +1,23 @@
+export const theme = {
+  grayLight: '#E0E0E0',
+  grayPrimary: '#9E9E9E',
+  grayDark: '#616161',
+  bluePrimary: '#2196F3',
+  blueDark: '#1976D2',
+  orangePrimary: '#FF9800',
+  orangeDark: '#F57C00',
+  tealPrimary: '#009688',
+  tealDark: '#00796B',
+};
+
+export const themeInverse = {
+  grayLight: '#616161',
+  grayPrimary: '#9E9E9E',
+  grayDark: '#E0E0E0',
+  bluePrimary: '#2196F3',
+  blueDark: '#1976D2',
+  orangePrimary: '#FF9800',
+  orangeDark: '#F57C00',
+  tealPrimary: '#009688',
+  tealDark: '#00796B',
+};
